requirements.txt
argon2-cffi==19.1.0
bcrypt==3.1.6
cffi==1.12.2
dj-database-url==0.5.0
Django==2.1.7
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-messages==0.5.3
django-pagedown==1.0.6
django-taggit==1.1.0
gunicorn==19.9.0
Pillow==5.4.1
psycopg2==2.8.1
pycparser==2.19
pytz==2018.9
six==1.12.0
whitenoise==4.1.2
